Add files via upload
This was created to ensure the OpenVPN server services are always running. This will prevent the case of the vpn server not being available after a server update occurs, or resources are too low. This script is set to run as a cron job. Extended usage is for Azure OMS, or CloudWatch queries to alert when service is not running.